Add Laurentian customizations for theses ETDMS output
To make LAC happy, we need an identifier linking to the PDF for a given
theses, the format of the thesis, and the level of the degree granted.

We're not currently using a qualified DC set of fields like
dc.degree.level, dc.degree.grantor, etc, so we have to rely on
parsing dc.description.degree to do the best we can.

Signed-off-by: Dan Scott <dscott@laurentian.ca>