Brutal hack to avoid DS-2623 (malformed error on upload)
Per the dspace-tech discussion with Kleiner Eike, actually including a
description in the field in which you are prompted to provide a
description results in a Malformed Error when you try to upload the
corresponding file. Apparently this is fixed in 6.0, but not in the
5.x releases. So instead, just remove that field entirely to avoid
people stumbling into the nasty bug.

Signed-off-by: Dan Scott <dscott@laurentian.ca>